#1 :wrench: search controller, autocomplete controller 분리 (1/2)
