+++ tiny_widgets/lib/controller/controllers.dart
@@ -1,1 +1,2 @@
-export 'search_bar_controller.dart';
+export 'search/search_controller.dart';
+export 'search/autocomplete_controller.dart';
